Split models.py into model_factories.py + auto-register all globals
- Extract COMPOSITE_NAME_SEPARATOR, GlobalModel, FavoriteModel, and all
  factory functions into api/model_factories.py. models.py now contains
  only auto-registration of globals and the hand-written core models
  (Piece, PieceState, UserProfile).

- Remove underscore prefix from factory functions — they are now the
  public contract of model_factories.py: make_simple_global_model,
  make_compose_global_models, make_favorite_model, dsl_field_to_django_field.

- Replace the manual per-global registration block in models.py with a
  _register_globals() loop that iterates get_global_names() and calls the
  appropriate factory for each entry, injecting results into the module
  namespace. Adding a new global to workflow.yml no longer requires any
  change to models.py.

- Add factory: false flag (workflow.yml + schema) for reference-only globals
  whose Django model is hand-written. Mark piece as factory: false — it is
  registered in globals for additional_fields DSL refs (@piece.name) but
  its model class is the hand-written Piece, not a GlobalModel subclass.

- Add is_factory_global() helper to api/workflow.py.

- Update test_model_factory.py to import factory helpers from
  api.model_factories rather than api.models.
