+++ api/tests/test_model_factory.py
@@ -1,6 +1,6 @@
 """Tests for the GlobalModel factory functions.
 
-Verifies that _make_simple_global_model and _make_compose_global_models
+Verifies that make_simple_global_model and make_compose_global_models
 generate classes with the correct fields, constraints, and helpers — without
 requiring DB access.  Migration-compatibility of the three converted simple
 globals (Location, ClayBody, GlazeMethod) is validated by running
@@ -11,15 +11,15 @@ from django.db import models
 
 
 # ---------------------------------------------------------------------------
-# _dsl_field_to_django_field
+# dsl_field_to_django_field
 # ---------------------------------------------------------------------------
 
 class TestDslFieldToDjangoField:
     """Unit tests for the DSL field → Django field mapping helper."""
 
     def _call(self, field_name: str, field_def: dict):
-        from api.models import _dsl_field_to_django_field
-        return _dsl_field_to_django_field(field_name, field_def)
+        from api.model_factories import dsl_field_to_django_field
+        return dsl_field_to_django_field(field_name, field_def)
 
     def test_name_field_is_char255_no_blank(self):
         f = self._call('name', {'type': 'string'})
@@ -71,17 +71,17 @@ class TestDslFieldToDjangoField:
 
 
 # ---------------------------------------------------------------------------
-# _make_simple_global_model — generated class structure
+# make_simple_global_model — generated class structure
 # ---------------------------------------------------------------------------
 
 class TestMakeSimpleGlobalModel:
-    """Verify the structure of classes generated by _make_simple_global_model."""
+    """Verify the structure of classes generated by make_simple_global_model."""
 
     @pytest.fixture
     def make(self, monkeypatch):
         """Return the factory, with _GLOBALS_MAP replaced by a minimal fixture."""
         import api.workflow as wf
-        import api.models as m
+        import api.model_factories as m
 
         fake_map = {
             'widget': {
@@ -113,7 +113,7 @@ class TestMakeSimpleGlobalModel:
             },
         }
         monkeypatch.setattr(wf, '_GLOBALS_MAP', fake_map)
-        return m._make_simple_global_model
+        return m.make_simple_global_model
 
     def test_class_name_matches_model_key(self, make):
         Widget = make('widget')
@@ -154,7 +154,7 @@ class TestMakeSimpleGlobalModel:
         """$ref fields in a global's fields section are additional_fields references
         and must not become model columns."""
         import api.workflow as wf
-        import api.models as m
+        import api.model_factories as m
         monkeypatch.setattr(wf, '_GLOBALS_MAP', {
             'thing': {
                 'model': 'Thing',
@@ -166,7 +166,7 @@ class TestMakeSimpleGlobalModel:
                 },
             },
         })
-        Thing = m._make_simple_global_model('thing')
+        Thing = m.make_simple_global_model('thing')
         field_names = {f.name for f in Thing._meta.get_fields()}
         assert 'name' in field_names
         assert 'clay_body' not in field_names
@@ -204,7 +204,7 @@ class TestMakeSimpleGlobalModel:
         """When the global config carries a 'plural' key the factory must use it
         for the user FK related_name instead of auto-pluralizing the global name."""
         import api.workflow as wf
-        import api.models as m
+        import api.model_factories as m
         monkeypatch.setattr(wf, '_GLOBALS_MAP', {
             'entity': {
                 'model': 'Entity',
@@ -214,7 +214,7 @@ class TestMakeSimpleGlobalModel:
                 'fields': {'name': {'type': 'string'}},
             },
         })
-        Entity = m._make_simple_global_model('entity')
+        Entity = m.make_simple_global_model('entity')
         user_field = Entity._meta.get_field('user')
         assert user_field.remote_field.get_accessor_name() == 'entities'
 
@@ -224,16 +224,16 @@ class TestMakeSimpleGlobalModel:
 
 
 # ---------------------------------------------------------------------------
-# _make_compose_global_models — generated class structure
+# make_compose_global_models — generated class structure
 # ---------------------------------------------------------------------------
 
 class TestMakeComposeGlobalModels:
-    """Verify the structure of classes generated by _make_compose_global_models."""
+    """Verify the structure of classes generated by make_compose_global_models."""
 
     @pytest.fixture
     def make(self, monkeypatch):
         import api.workflow as wf
-        import api.models as m
+        import api.model_factories as m
 
         fake_map = {
             'component': {
@@ -269,7 +269,7 @@ class TestMakeComposeGlobalModels:
             },
         }
         monkeypatch.setattr(wf, '_GLOBALS_MAP', fake_map)
-        return m._make_compose_global_models
+        return m.make_compose_global_models
 
     def test_returns_two_model_classes(self, make):
         CompositeModel, ThroughModel = make('composite')
@@ -349,26 +349,26 @@ class TestMakeComposeGlobalModels:
 
     def test_no_compose_from_raises(self, monkeypatch):
         import api.workflow as wf
-        import api.models as m
+        import api.model_factories as m
         monkeypatch.setattr(wf, '_GLOBALS_MAP', {
             'plain': {'model': 'Plain', 'public': False, 'private': True,
                       'fields': {'name': {'type': 'string'}}},
         })
         with pytest.raises(ValueError, match='no compose_from'):
-            m._make_compose_global_models('plain')
+            m.make_compose_global_models('plain')
 
 
 # ---------------------------------------------------------------------------
-# _make_favorite_model — generated class structure
+# make_favorite_model — generated class structure
 # ---------------------------------------------------------------------------
 
 class TestMakeFavoriteModel:
-    """Verify the structure of classes generated by _make_favorite_model."""
+    """Verify the structure of classes generated by make_favorite_model."""
 
     @pytest.fixture
     def make(self, monkeypatch):
         import api.workflow as wf
-        import api.models as m
+        import api.model_factories as m
 
         fake_map = {
             'widget': {
@@ -380,7 +380,7 @@ class TestMakeFavoriteModel:
             },
         }
         monkeypatch.setattr(wf, '_GLOBALS_MAP', fake_map)
-        return m._make_favorite_model
+        return m.make_favorite_model
 
     def test_class_name_is_favorite_model_name(self, make):
         FavWidget = make('widget')
@@ -427,7 +427,7 @@ class TestMakeFavoriteModel:
     def test_plural_key_used_in_user_related_name(self, monkeypatch):
         """explicit plural key overrides auto-pluralize for related_name."""
         import api.workflow as wf
-        import api.models as m
+        import api.model_factories as m
         monkeypatch.setattr(wf, '_GLOBALS_MAP', {
             'entity': {
                 'model': 'Entity',
@@ -438,7 +438,7 @@ class TestMakeFavoriteModel:
                 'fields': {'name': {'type': 'string'}},
             },
         })
-        FavEntity = m._make_favorite_model('entity')
+        FavEntity = m.make_favorite_model('entity')
         user_field = FavEntity._meta.get_field('user')
         assert user_field.remote_field.get_accessor_name() == 'favorite_entities'
