+++ workflow.schema.yml
@@ -70,6 +70,16 @@ $defs:
           Whether this global type supports user-private instances.
           Private objects are owned by a specific user and invisible to others.
         default: true
+      factory:
+        type: boolean
+        description: |
+          Whether the model factory should generate a Django model class for this
+          global. Defaults to true. Set to false for "reference-only" globals whose
+          model is hand-written elsewhere (e.g. the core Piece model, which is
+          registered in globals only so that additional_fields DSL refs like
+          "@piece.name" can be validated, but whose model class is not a GlobalModel
+          subclass and must not be auto-generated).
+        default: true
       favoritable:
         type: boolean
         description: |
